requirements.txt
asgiref==3.4.1
Django==3.2.7
django-taggit==1.5.1
Markdown==3.3.4
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
